requirements.txt
p5==0.6.0
absl-py==0.9.0